refactor: removed defensive programming paths that had no coverage
