+++ .env.sample
@@ -7,9 +7,6 @@
 
 # DEBUG="*"
 
-# OPTIONAL: Token to publish GitHub packages.
-NPM_TOKEN=
-
 # Harbor task
 HARBOR_USER='admin'
 HARBOR_PASSWORD=''
